Refactor to new test module
